try a new algorithm

+++ src/module/spell/spell-internal.h
@@ -44,7 +44,7 @@ typedef struct {
 
 typedef struct {
     char *word;
-    float dist;
+    int dist;
 } SpellCustomCWord;
 
 typedef struct {
